Clearer subfolder specification for .gitignore
In some Unity projects there are two Package folders. The one on the top level of the project should be tracked as it contains the manifest.json package dependencies. The one in the Assets folder should be ignored as it contains .dlls e.g. from a local Bonsai environment that should be populated when building the project.