Update gitignore to specifically ignore TMP assets
TMP in unity does a funny thing where it is part package, part dialog import. In the end this means that TMP resources get added directly to assets when imported, but since if they are not tracked in the repo unity pops up a dialog asking you to import we don't need to have a copy of them in the tracked repo.